test(ssa-pipeline): move seed tests into seed module
