+++ crates/ssa-pipeline/src/stochastic/seed.rs
@@ -151,3 +151,158 @@ impl<const N: usize> AsMut<[Xoshiro256PlusPlus; N]> for RngStreams<N> {
         &mut self.rngs
     }
 }
+
+#[cfg(test)]
+#[cfg_attr(coverage_nightly, coverage(off))]
+mod tests {
+    use rand::Rng;
+
+    use super::*;
+
+    const TEST_MODEL: SimulationModel = SimulationModel::new("experiment/ssa-pipeline-test/v1");
+    const MAIN_VARIABLE: RandomVariable = RandomVariable::new("ssa/main/v1");
+    const SEGREGATION_VARIABLE: RandomVariable = RandomVariable::new("model/segregation/v1");
+    const MUTATION_VARIABLE: RandomVariable = RandomVariable::new("model/mutation/v1");
+
+    #[test]
+    fn test_model_and_random_variable_names_are_stable() {
+        let model = SimulationModel::new("experiment/test/v1");
+        let variable = RandomVariable::new("test/stream/v1");
+
+        assert_eq!(model.as_str(), "experiment/test/v1");
+        assert_eq!(variable.as_str(), "test/stream/v1");
+        assert_eq!(model.to_string(), "experiment/test/v1");
+        assert_eq!(variable.to_string(), "test/stream/v1");
+    }
+
+    #[test]
+    fn test_stream_seed_debug_output_is_redacted() {
+        let stream_seed = TEST_MODEL.derive_stream_seed(MAIN_VARIABLE);
+
+        let stream_debug = format!("{stream_seed:?}");
+
+        assert_eq!(stream_debug, "StreamSeed { .. }");
+        assert!(!stream_debug.contains("bytes"));
+    }
+
+    #[test]
+    fn test_stream_seed_streams_are_stable_and_isolated() {
+        let segregation_seed = TEST_MODEL.derive_stream_seed(SEGREGATION_VARIABLE);
+        let mutation_seed = TEST_MODEL.derive_stream_seed(MUTATION_VARIABLE);
+        let mut segregation1 = segregation_seed.make_stream(b"input-A");
+        let mut segregation2 = segregation_seed.make_stream(b"input-A");
+        let mut mutation = mutation_seed.make_stream(b"input-A");
+
+        assert_eq!(segregation1.next_u64(), segregation2.next_u64());
+        assert_ne!(segregation1.next_u64(), mutation.next_u64());
+    }
+
+    #[test]
+    fn test_stream_seed_input_bytes_change_stream() {
+        let seed = TEST_MODEL.derive_stream_seed(SEGREGATION_VARIABLE);
+        let mut rng_a = seed.make_stream(b"input-A");
+        let mut rng_b = seed.make_stream(b"input-B");
+
+        assert_ne!(rng_a.next_u64(), rng_b.next_u64());
+    }
+
+    #[test]
+    fn test_stream_seed_material_boundaries_change_stream() {
+        let seed_ab_c = SimulationModel::new("ab").derive_stream_seed(RandomVariable::new("c"));
+        let seed_a_bc = SimulationModel::new("a").derive_stream_seed(RandomVariable::new("bc"));
+        let mut rng_ab_c = seed_ab_c.make_stream(b"input-A");
+        let mut rng_a_bc = seed_a_bc.make_stream(b"input-A");
+
+        assert_ne!(rng_ab_c.next_u64(), rng_a_bc.next_u64());
+    }
+
+    #[test]
+    fn test_single_stream_seed_known_sequence() {
+        let seed = TEST_MODEL.derive_single_stream_seed();
+        let mut rng = seed.make_stream(b"input-A");
+
+        assert_eq!(rng.next_u64(), 18_439_914_891_317_024_306);
+        assert_eq!(rng.next_u64(), 3_187_733_765_017_413_158);
+    }
+
+    #[test]
+    fn test_stream_seed_known_sequence() {
+        let seed = TEST_MODEL.derive_stream_seed(SEGREGATION_VARIABLE);
+        let mut rng = seed.make_stream(b"input-A");
+
+        assert_eq!(rng.next_u64(), 14_685_298_993_141_689_080);
+        assert_eq!(rng.next_u64(), 5_115_129_566_991_615_718);
+    }
+
+    #[test]
+    fn test_stream_seed_bundle_allows_duplicate_variables() {
+        let seeds = TEST_MODEL.derive_stream_seeds([SEGREGATION_VARIABLE, SEGREGATION_VARIABLE]);
+        let mut streams = seeds.make_streams(b"input-A");
+        let [left, right] = streams.as_mut();
+
+        assert_eq!(left.next_u64(), right.next_u64());
+    }
+
+    #[test]
+    fn test_stream_seed_bundle_accessors_preserve_order() {
+        let seeds = TEST_MODEL.derive_stream_seeds([SEGREGATION_VARIABLE, MUTATION_VARIABLE]);
+        let [segregation_seed_ref, mutation_seed_ref] = seeds.as_ref();
+        let mut segregation_rng = segregation_seed_ref.make_stream(b"input-A");
+        let mut mutation_rng = mutation_seed_ref.make_stream(b"input-A");
+
+        let [segregation_seed, mutation_seed] = seeds.into_inner();
+        let mut owned_segregation_rng = segregation_seed.make_stream(b"input-A");
+        let mut owned_mutation_rng = mutation_seed.make_stream(b"input-A");
+
+        assert_eq!(segregation_rng.next_u64(), owned_segregation_rng.next_u64());
+        assert_eq!(mutation_rng.next_u64(), owned_mutation_rng.next_u64());
+    }
+
+    #[test]
+    fn test_single_stream_seed_is_isolated_from_named_stream_seed() {
+        let single_seed = TEST_MODEL.derive_single_stream_seed();
+        let stream_seed = TEST_MODEL.derive_stream_seed(MAIN_VARIABLE);
+        let mut single_rng = single_seed.make_stream(b"input-A");
+        let mut stream_rng = stream_seed.make_stream(b"input-A");
+
+        assert_ne!(single_rng.next_u64(), stream_rng.next_u64());
+    }
+
+    #[test]
+    fn test_single_stream_seed_matches_empty_random_variable() {
+        let single_seed = TEST_MODEL.derive_single_stream_seed();
+        let empty_variable_seed = TEST_MODEL.derive_stream_seed(RandomVariable::new(""));
+        let mut single_rng = single_seed.make_stream(b"input-A");
+        let mut empty_variable_rng = empty_variable_seed.make_stream(b"input-A");
+
+        assert_eq!(single_rng.next_u64(), empty_variable_rng.next_u64());
+    }
+
+    #[test]
+    fn test_stream_seed_bundle_supports_multiple_mutable_rngs() {
+        let seeds = TEST_MODEL.derive_stream_seeds([SEGREGATION_VARIABLE, MUTATION_VARIABLE]);
+        let mut streams = seeds.make_streams(b"input-A");
+        let [segregation_rng, mutation_rng] = streams.as_mut();
+
+        let segregation_value = segregation_rng.next_u64();
+        let mutation_value = mutation_rng.next_u64();
+
+        assert_ne!(segregation_value, mutation_value);
+    }
+
+    #[test]
+    fn test_rng_stream_bundle_into_inner_preserves_order() {
+        let seeds = TEST_MODEL.derive_stream_seeds([SEGREGATION_VARIABLE, MUTATION_VARIABLE]);
+        let mut streams = seeds.make_streams(b"input-A");
+        let [segregation_rng, mutation_rng] = streams.as_mut();
+
+        let segregation_value = segregation_rng.next_u64();
+        let mutation_value = mutation_rng.next_u64();
+
+        let streams = seeds.make_streams(b"input-A");
+        let [mut owned_segregation_rng, mut owned_mutation_rng] = streams.into_inner();
+
+        assert_eq!(segregation_value, owned_segregation_rng.next_u64());
+        assert_eq!(mutation_value, owned_mutation_rng.next_u64());
+    }
+}
